+++ res/views/components/error.blade.php
@@ -1,0 +1,4 @@
+@error($field, $stack)
+    @ui($element, $attributes, $message)
+    {{ $slot }}
+@enderror