+++ res/views/components/errors.blade.php
@@ -1,0 +1,8 @@
+@if ($errors->any())
+    @ui($element, $attributes)
+        @foreach ($errors->all() as $error)
+            @ui($errorElement, $errorAttributes, $error)
+        @endforeach
+        {{ $slot }}
+    @endui
+@endif