+++ res/views/samples/form.blade.php
@@ -1,0 +1,5 @@
+<s-errors/>
+
+<x-ui::form>
+    <s-error field="name"/>
+</x-ui::form>